pyproject.toml
[project]
name = "smartestiroid"
version = "0.1.0"
description = "AI-powered Android app testing framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.0",
    "anyio>=4.10.0",
    "appium-python-client>=5.2.4",
    "asyncio>=4.0.0",
    "colorama>=0.4.6",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "typing-extensions>=4.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smartestiroid"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""


